Added helper script repo_dataverse.py for file transfer with Dataverse repository. Added helper script repodown.sh to download and decompress file from data repo. Adding container building github action. Minor bug correction. Minor code improvement. Documentation update.

+++ doc/scripts/install.sh
@@ -2,7 +2,7 @@
 
 function usage()
 {
-	fmt='%-25s%s\n'
+	fmt='  %-25s%s\n'
 	echo "Usage: $(basename "$0") [-h]" >&2
 	echo "Installing Dictys as a conda environment" >&2
 	echo "" >&2	
@@ -16,6 +16,15 @@ function usage()
 	printf "$fmt" '' 'You need a version supported by your driver and by pytorch.' >&2
 	printf "$fmt" '' 'See https://stackoverflow.com/questions/9727688/how-to-get-the-cuda-version' >&2
 	printf "$fmt" 'COMMIT_VERSION' 'Dictys commit to install. Accepts hash, branch name, and tag. Defaults to master.' >&2
+	echo '' >&2
+	echo "Advanced options in environmental variables" >&2
+	printf "$fmt" 'LOCAL_VERSION' 'Use a local folder to intall Dictys from. If specified, does not download from repo and ignores COMMIT_VERSION.' >&2
+	printf "$fmt" 'STEPS' 'Binary keys for which installation steps to perform. Accepts:' >&2
+	printf "$fmt" '' '1: conda steps' >&2
+	printf "$fmt" '' '2: pip steps' >&2
+	printf "$fmt" '' '4: homer update' >&2
+	printf "$fmt" '' 'Default: all steps' >&2
+	printf "$fmt" 'PIP_OPTIONS' 'Options for pip install' >&2
 	exit 1
 }
 
@@ -45,20 +54,32 @@ else
 	conda_deps="cudatoolkit=$CUDAVERSION_CONDA"
 fi
 
-#Install non-pypi dependencies: pytorch, bedtools, homer, samtools, macs2, ffmpeg
-conda create -y -n $CONDAENV_NAME -c bioconda -c conda-forge -c pytorch python=$PYTHONVERSION_CONDA pytorch torchvision torchaudio $conda_deps bedtools homer samtools macs2 ffmpeg
-#You may need "conda activate ..." instead
-. activate $CONDAENV_NAME
-#Install pypi dependencies
-pip install numpy pandas docutils h5py pyro-ppl==1.6.0 scipy networkx pybedtools pyDNase threadpoolctl joblib matplotlib jupyter adjustText
-#Install Dictys
-pip install git+https://github.com/pinellolab/dictys.git@$COMMIT_VERSION
-#Correcting matplotlib version due to pyDNase dependency
-pip uninstall -y pyDNase
-pip install -U matplotlib
-pip install --no-deps pyDNase
-#Update homer
-cd "$(dirname "$(dirname "$(realpath "$(which homer)")")")"
-./configureHomer.pl -update
-chmod u+x configureHomer.pl
-conda deactivate
+if [ "a$STEPS" == "a" ] || [ "a$(( STEPS & 1 ))" != "a0" ]; then
+	#Install non-pypi dependencies: pytorch, bedtools, homer, samtools, macs2, ffmpeg
+	conda create -y -n $CONDAENV_NAME -c bioconda -c conda-forge -c pytorch python=$PYTHONVERSION_CONDA pytorch torchvision torchaudio $conda_deps bedtools homer samtools macs2 ffmpeg
+	#You may need "conda activate ..." instead
+	. activate $CONDAENV_NAME
+fi
+if [ "a$STEPS" == "a" ] || [ "a$(( STEPS & 2 ))" != "a0" ]; then
+	#Install Dictys
+	if [ "a$LOCAL_VERSION" == "a" ]; then
+		pip install $PIP_OPTIONS git+https://github.com/pinellolab/dictys.git@$COMMIT_VERSION
+	else
+		pip install $PIP_OPTIONS "$LOCAL_VERSION"
+	fi
+	#Correcting matplotlib version due to pyDNase dependency
+	pip uninstall -y pyDNase
+	pip install -U matplotlib
+	pip install $PIP_OPTIONS --no-deps pyDNase
+fi
+if [ "a$STEPS" == "a" ] || [ "a$(( STEPS & 4 ))" != "a0" ]; then
+	#Update homer
+	pushd . &> /dev/null
+	cd "$(dirname "$(dirname "$(realpath "$(which homer)")")")"
+	./configureHomer.pl -update
+	chmod u+x configureHomer.pl
+	popd &> /dev/null
+fi
+if [ "a$STEPS" == "a" ] || [ "a$(( STEPS & 1 ))" != "a0" ]; then
+	conda deactivate
+fi
